@@ -1,34 +1,37 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Description>A simplified mapper that also works with immutable types.</Description>
-    <PackageLicenseFile>LICENSE</PackageLicenseFile>
-    <PackageTags>Automapper, Mapper, Map, IMap, Immutable, Attribute Map</PackageTags>
-    <Copyright>@2022</Copyright>
-    <PackageProjectUrl>https://github.com/MichaelPuckett2/ImmutaMap</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/MichaelPuckett2/ImmutaMap</RepositoryUrl>
-    <Company>Triple G3</Company>
-    <Authors>Michael Puckett, II</Authors>
-    <Version>6.0.0</Version>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net7.0</TargetFramework>
+		<Nullable>enable</Nullable>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<WarningsAsErrors>Nullable</WarningsAsErrors>
+		<Description>A simplified mapper that also works with immutable types.</Description>
+		<PackageLicenseFile>LICENSE</PackageLicenseFile>
+		<PackageTags>Automapper, Mapper, Map, IMap, Immutable, Attribute Map</PackageTags>
+		<Copyright>@2022</Copyright>
+		<PackageProjectUrl>https://github.com/MichaelPuckett2/ImmutaMap</PackageProjectUrl>
+		<RepositoryUrl>https://github.com/MichaelPuckett2/ImmutaMap</RepositoryUrl>
+		<Company>Triple G3</Company>
+		<Authors>Michael Puckett, II</Authors>
+		<Version>7.0.0</Version>
+		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <Compile Remove="ReflectionTools\**" />
-    <EmbeddedResource Remove="ReflectionTools\**" />
-    <None Remove="ReflectionTools\**" />
-  </ItemGroup>
+	<ItemGroup>
+		<Compile Remove="ReflectionTools\**" />
+		<EmbeddedResource Remove="ReflectionTools\**" />
+		<None Remove="ReflectionTools\**" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <None Include="..\LICENSE">
-      <Pack>True</Pack>
-      <PackagePath></PackagePath>
-    </None>
-  </ItemGroup>
+	<ItemGroup>
+		<None Include="..\LICENSE">
+			<Pack>True</Pack>
+			<PackagePath></PackagePath>
+		</None>
+	</ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
+	</ItemGroup>
 
 </Project>
