@@ -1,9 +1,9 @@
 ﻿using System.Reflection;
 
-namespace ImmutaMap.Interfaces
+namespace ImmutaMap.Interfaces;
+
+public interface IMapping
 {
-    public interface IMapping
-    {
-        bool TryGetValue<TSource>(TSource source, PropertyInfo sourcePropertyInfo, PropertyInfo targetPropertyInfo, object previouslyMappedValue, out object result);
-    }
+    bool TryGetValue<TSource>(TSource source, PropertyInfo sourcePropertyInfo, PropertyInfo targetPropertyInfo, out object result);
+    bool TryGetValue<TSource>(TSource source, PropertyInfo sourcePropertyInfo, PropertyInfo targetPropertyInfo, object previouslyMappedValue, out object result);
 }
