@@ -1,43 +1,40 @@
 ﻿using ImmutaMap.Interfaces;
-using System.Collections.Generic;
-using System.Linq;
 
-namespace ImmutaMap
+namespace ImmutaMap;
+
+public class Map<TSource, TTarget>
 {
-    public class Map<TSource, TTarget>
+    private readonly List<(string SourceProperty, string ResultProperty)> propertyNameMaps = new();
+    private readonly IList<IMapping> mappings = new List<IMapping>();
+
+    public Map(TSource source, MapConfiguration<TTarget> mapConfiguration)
     {
-        private readonly List<(string SourceProperty, string ResultProperty)> propertyNameMaps = new List<(string SourceProperty, string ResultProperty)>();
-        private readonly IList<IMapping> mappings = new List<IMapping>();
+        Source = source;
+        MapConfiguration = mapConfiguration;
+        IgnoringCase = mapConfiguration.IgnoreCase;
+        WillNotThrowExceptions = mapConfiguration.WillNotThrowExceptions;
+    }
 
-        public Map(bool ignoreCase = false, bool throwExceptions = true) : this(default, ignoreCase, throwExceptions) { }
-        public Map(TSource source, bool ignoreCase = false, bool throwExceptions = true)
-        {
-            Source = source;
-            IsIgnoringCase = ignoreCase;
-            IsThrowingExceptions = throwExceptions;
-        }
+    public void AddMapping(IMapping mapping) => mappings.Add(mapping);
+    internal IEnumerable<IMapping> Mappings => mappings;
 
-        public void AddMapping(IMapping mapping) => mappings.Add(mapping);
-        internal IEnumerable<IMapping> Mappings => mappings;
+    internal IEnumerable<(string SourceProperty, string ResultProperty)> PropertyNameMaps => propertyNameMaps.ToList();
+    internal MapConfiguration<TTarget> MapConfiguration { get; }
 
-        internal IEnumerable<(string SourceProperty, string ResultProperty)> PropertyNameMaps => propertyNameMaps.ToList();
+    public TSource Source { get; internal set; }
 
-        public TSource Source { get; internal set; }
+    /// <summary>
+    /// When true the MapBuilder will ignore property case sensitivity while mapping the types.
+    /// </summary>
+    public bool IgnoringCase { get; }
 
-        /// <summary>
-        /// When true the MapBuilder will ignore property case sensitivity while mapping the types.
-        /// </summary>
-        public bool IsIgnoringCase { get; }
+    /// <summary>
+    /// When true mapping will throw exceptions on properties that cannot be mapped, otherwise they are silently skipped and mapping continues.
+    /// </summary>
+    public bool WillNotThrowExceptions { get; }
 
-        /// <summary>
-        /// Is set to true the MapBuilder will ignore exceptions, and map only the properties that it can.
-        /// If false, MapBuilder will throw an exception if the property cannot be mapped.
-        /// </summary>
-        public bool IsThrowingExceptions { get; }
-
-        public void AddPropertyNameMapping(string sourcePropertyName, string targetPropertyName)
-        {
-            propertyNameMaps.Add((sourcePropertyName, targetPropertyName));
-        }
+    public void AddPropertyNameMapping(string sourcePropertyName, string targetPropertyName)
+    {
+        propertyNameMaps.Add((sourcePropertyName, targetPropertyName));
     }
 }
